+++ src/map/packets/s2c/0x032_event.cpp
@@ -39,9 +39,9 @@ GP_SERV_COMMAND_EVENT::GP_SERV_COMMAND_EVENT(const CCharEntity* PChar, const Eve
         packet.ActIndex = PChar->targid;
     }
 
-    packet.EventNum   = PChar->getZone();
+    packet.EventNum   = static_cast<uint16>(PChar->getZone());
     packet.EventPara  = eventInfo->eventId;
     packet.Mode       = eventInfo->eventFlags & 0xFFFF;
-    packet.EventNum2  = PChar->getZone();
+    packet.EventNum2  = static_cast<uint16>(PChar->getZone());
     packet.EventPara2 = eventInfo->eventFlags >> 16;
 }
